+++ src/App.tsx
@@ -38,7 +38,12 @@ export default function TheMemorialDay() {
       </Route>
       <Route path={ST_PATH} element={<MainLayout />} />
       <Route path={HO_PATH} element={<MainLayout />} />
-      <Route path={SU_PATH} element={<Support />} />
+      <Route path={SU_PATH} element={<MainLayout />} >
+        <Route path={SU_PATH} element={<Support />} />
+      </Route>
+      
+
+
       <Route path={JO_PATH} element={<MainLayout />} />
       <Route path={OTHERS_PATH} element={<Index />} />
     </Routes>
