--- src/view/Support/Notice/style.css
@@ -1,78 +1,0 @@
-#notice-wrapper {
-    display: flex;
-    flex-direction: column;
-    gap: 30px;
-}
-
-#notice-wrapper .title {
-    text-align: center;
-    font-size: 30px;
-}
-
-#notice-wrapper .navigation {
-    display: flex;
-    justify-content: center;
-    gap: 60px;
-
-    font-size: 20px;
-    cursor: pointer;
-}
-
-#notice-wrapper .top-text {
-    font-size: 18px;
-}
-
-#notice-wrapper .table {
-    display: flex;
-    flex-direction: column;
-
-    border-top: 2px solid rgba(0, 0, 0, 1);
-    border-bottom: 2px solid rgba(0, 0, 0, 1);
-}
-
-#notice-wrapper .table .th {
-    display: flex;
-    justify-content: space-between;
-
-    border-bottom: 2px solid rgba(0, 0, 0, 1);
-
-    font-size: 18px;
-    font-weight: 700;
-    
-    padding: 0 50px;
-}
-
-#notice-wrapper .table .tr {
-    display: flex;
-    gap: 40px;
-
-    padding: 0px 25px;
-}
-
-#notice-wrapper .table .tr .td-no {
-    width: 70px;
-    text-align: center;
-}
-
-#notice-wrapper .table .tr .td-title {
-    flex: 1;
-}
-
-#notice-wrapper .table .tr .td-date {
-    width: 130px;
-    text-align: center;
-}
-
-#notice-wrapper .bottom {
-    position: relative;
-    display: flex;
-    justify-content: center;
-}
-
-#notice-wrapper .search-box {
-    position: absolute;
-    right: 0;
-
-    display: flex;
-    justify-content: end;
-}
